move tests to single directory + make tests work on local environment
