--- src/tests/e2e/example.js
@@ -1,13 +1,0 @@
-module.exports = {
-  'Demo test Google' : function (browser) {
-    browser
-      .url('https://www.google.com')
-      .waitForElementVisible('body')
-      .setValue('input[type=text]', 'nightwatch')
-      .waitForElementVisible('input[name=btnK]')
-      .click('input[name=btnK]')
-      .pause(1000)
-      .assert.containsText('#main', 'Night Watch')
-      .end();
-  }
-};
