+++ webpack.common.js
@@ -45,7 +45,7 @@ module.exports = {
     new HtmlWebpackRootPlugin("app")
   ],
   output: {
-    filename: '[name].[contenthash].js',
+    filename: '[name].[hash].js',
     path: path.resolve(__dirname, "dist")
   },
   optimization: {
